- move fromTag resolving to main.ts and remove dependency on checks from ReleaseNotes.ts   - require fromTag for ReleaseNotes - adjust tests, split into according test classes - simplify transform.ts
